+++ lib/rubocop-rspec_rails.rb
@@ -21,7 +21,7 @@ RuboCop::ConfigLoader.inject_defaults!(project_root)
 # https://github.com/rubocop/rubocop-rspec_rails/issues/8
 module RuboCop
   module Cop
-    class AmbiguousCopName # rubocop:disable Style/Documentation
+    class Registry # rubocop:disable Style/Documentation
       prepend(Module.new do
         def qualified_cop_name(name, path, warn: true)
           return super unless name == 'RSpec/Rails/HttpStatus'